refactor model_configurations and model_configuration to model_configuration_helper
