+++ classes/model_configuration_helper.php
@@ -30,7 +30,7 @@ use core_analytics\manager;
 /**
  * Class for the list of model configurations.
  */
-class model_configurations {
+class model_configuration_helper {
 
     /**
      * Collect all model configuration objects.
@@ -79,7 +79,7 @@ class model_configurations {
                     }
 
                     if (!$modelalreadyhasexactsameconfig) {
-                        $modelconfigids[] = model_configuration::create_and_get_for_model($modelid);
+                        $modelconfigids[] = self::create_and_get_for_model($modelid);
                     }
                 }
             }
@@ -92,7 +92,7 @@ class model_configurations {
             $target = manager::get_target($targetname);
             if (!$target->based_on_assumptions()) {
                 // For now, ignore static models and only handle machine learning models.
-                $modelconfigids[] = model_configuration::create_and_get_for_model($missingmodelid);
+                $modelconfigids[] = self::create_and_get_for_model($missingmodelid);
             }
         }
 
@@ -114,5 +114,64 @@ class model_configurations {
     private static function get_settings_values(stdClass $settingdbentry): array {
         $vals = array_values((array) $settingdbentry);
         return array_slice($vals, 1, null);
+    }
+
+    /**
+     * Helper method: Short check to verify whether the provided value is valid, and thus a valid list exists.
+     *
+     * @param string|null $value to check
+     * @return boolean
+     */
+    public static function valid_exists(?string $value): bool {
+        return isset($value) && $value != "" && $value != "[]";
+    }
+
+    /**
+     * Create a new model configuration for a model id.
+     * Failing gracefully: If config for this model already exists, just return it.
+     * The accompanying db table is needed to preserve a record of the model configuration
+     * even if the model has been deleted.
+     *
+     * @param int $modelid of an analytics model
+     * @return int id of created or retrieved object
+     */
+    public static function create_and_get_for_model(int $modelid): int {
+        global $DB;
+
+        $modelobj = $DB->get_record('analytics_models', ['id' => $modelid], '*', MUST_EXIST);
+
+        $obj = new stdClass();
+        $obj->modelid = $modelid;
+        $obj->name = $modelobj->name;
+        if (!isset($obj->name)) {
+            $modelidcount = $DB->count_records('tool_lala_model_configs', ['modelid' => $modelid]);
+            $obj->name = 'config' . $modelid . '/' . $modelidcount;
+        }
+        $obj->target = $modelobj->target;
+
+        if (self::valid_exists($modelobj->predictionsprocessor)) {
+            $obj->predictionsprocessor = $modelobj->predictionsprocessor;
+        } else {
+            $default = manager::default_mlbackend();
+            $obj->predictionsprocessor = $default;
+        }
+
+        if (self::valid_exists($modelobj->timesplitting)) {
+            $obj->analysisinterval = $modelobj->timesplitting;
+        } else {
+            $analysisintervals = manager::get_time_splitting_methods_for_evaluation();
+            $firstanalysisinterval = array_keys($analysisintervals)[0];
+            $obj->analysisinterval = $firstanalysisinterval;
+        }
+
+        if (self::valid_exists($modelobj->contextids)) {
+            $obj->defaultcontextids = $modelobj->contextids;
+        }
+
+        $obj->indicators = $modelobj->indicators;
+
+        $obj->timecreated = time();
+
+        return $DB->insert_record('tool_lala_model_configs', $obj);
     }
 }
