+++ classes/output/renderer.php
@@ -78,8 +78,8 @@ class renderer extends plugin_renderer_base {
      * @param model_configurations $page
      * @return string html for the page
      */
-    public function render_model_configurations(model_configurations $page): string {
+    public function render_model_configuration_helper(model_configurations $page): string {
         $data = $page->export_for_template($this);
-        return parent::render_from_template('tool_lala/model_configurations', $data);
+        return parent::render_from_template('tool_lala/model_configuration_helper', $data);
     }
 }
